Code fixex + Code cleanup
fixed get by email
added get all users
