+++ src/main/resources/application.properties
@@ -1,6 +1,16 @@
+
 spring.application.name=booking-java-bootcamp
 
 springdoc.swagger-ui.enabled=true
 springdoc.swagger-ui.path=/swagger.html
-// enabling CSRF support, while using standard header names
+# enabling CSRF support, while using standard header names
 springdoc.swagger-ui.csrf.enabled=true
+
+# enabling access logs
+server.tomcat.accesslog.enabled=true
+server.tomcat.accesslog.suffix=.log
+server.tomcat.accesslog.prefix=access_log
+server.tomcat.accesslog.file-date-format=.yyyy-MM-dd
+server.tomcat.basedir=tomcat
+server.tomcat.accesslog.directory=logs
+server.tomcat.accesslog.pattern=common
